third_party/blink/web_tests/platform/linux-chrome/external/wpt/html/semantics/interactive-elements/the-dialog-element/focus-previous-iframe.tentative-expected.txt

This is a testharness.js-based test.
Harness Error. harness_status.status = 1 , harness_status.message = Uncaught Error: assert_equals: expected Element node <input></input> but got Element node <iframe srcdoc="<input><dialog> Dialog in child </dialog>...
[PASS] Focus should move back from parent document to child document
[PASS] Focus should move back from child document to parent document
Harness: the test ran to completion.
